synchronize the code with the commercial excel plugin

+++ engine/org.eclipse.birt.report.engine.emitter.prototype.excel/src/org/eclipse/birt/report/engine/emitter/excel/ExcelUtil.java
@@ -558,12 +558,12 @@ public class ExcelUtil
 	}
 
 	// change the columnWidth unit from point to characterNumber
-	public static double convertColWidth( float width )
+	public static double convertColWidth( double width )
 	{
 		return convertColWidth( width, 96 );
 	}
 
-	public static double convertColWidth( float width, int dpi )
+	public static double convertColWidth( double width, int dpi )
 	{
 		float PX_PT = INCH_PT / dpi;
 		// TODO: more study about the caculation
@@ -1274,7 +1274,7 @@ public class ExcelUtil
 	{
 		// Make sure the name you entered does not exceed 31 characters.
 		// Make sure the name does not contain any of the following characters:
-		// \ / ? *: [ or ]
+		// \ / ? * : [ or ]
 		name = name.replaceAll( "[\\\\/?*:\\[\\] ]", "_" );
 		if ( name.length( ) > SHEETNAME_LENGTH )
 		{
